fix: address additional CodeRabbit feedback (round 3) on #76
- Split BriefHandlers.cs into one file per use case (vertical slice convention).
- EditPendingBriefUpdate: reject invalid risk severities with ArgumentException
  instead of silently coercing to Medium (matches RaiseInitiativeBriefRisk).
- ApplyProposalToInitiative: pre-check risk existence/status before ResolveRisk
  so terminal-state ArgumentExceptions propagate instead of being swallowed.
- BriefRefreshQueue.Enqueue: replace racy `_inFlight[key] = 1` indexer write with
  TryUpdate loop so MarkComplete's TryRemove cannot leave a ghost rerun flag.
- living-brief-tab: reload brief after refreshNow() succeeds so auto-applied
  updates are visible in the UI immediately.
